chore: operf refactoring
